bioreasonc-agent/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
anthropic>=0.18.0
neo4j>=5.14.0
pyyaml>=6.0
notebook>=7.0.0
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0
